instance/requirements.txt
bottle
psutil
mcrcon
python-dotenv